requirements.txt
fastapi
uvicorn
pillow<=11.0
pydantic
torchvision
python-multipart
# pip install --pre -U --extra-index-url https://storage.openvinotoolkit.org/simple/wheels/pre-release openvino_tokenizers openvino
# pip install git+https://github.com/huggingface/optimum-intel.git
